Use upstream LoRA mapping type
