+++ vllm_metal/v1/lora/__init__.py
@@ -1,7 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
+from vllm.lora.layers import LoRAMapping
+
 from .layers import MLXLinearWithLoRA, can_wrap
-from .mapping import LoRAMapping, LoRAMappingBuilder
 from .model_manager import MLXLoRAModelManager
 from .peft_loader import LoadedLoRA, LoRALayerWeightsMLX, load_peft_adapter
 from .punica_wrapper import PunicaWrapperMLX
@@ -10,7 +11,6 @@ from .worker_manager import MetalWorkerLoRAManager
 
 __all__ = [
     "LoRAMapping",
-    "LoRAMappingBuilder",
     "PunicaWrapperMLX",
     "MLXLinearWithLoRA",
     "can_wrap",
